assets:尝试摄取食物相关

+++ Assets/Scripts/Gameplay/JobSystem/JobTracker.cs
@@ -11,7 +11,7 @@ public class ThingUnit_JobTracker
 {
     private const int THINK_INTERVAL = 30;
 
-    private int LastTimeThinkTick;
+    private long LastTimeThinkTick;
 
     public Thing_Unit Unit;
